feat(login): Safer login cookie
The current login cookie implementation uses the hash of the user's password, as stored in the database, to log a user in. That allows an attacker with a copy of the database to log into a user's account without knowing their password. Since many people re-use their password across multiple sites, this allows an attacker to log into users' accounts on all those sites.

This new implementation generates a random key on every login. A hashed version of that key is stored in the database, along with the ID of the user to whom that key belongs and an expiration time. Because a hashed version of the cookie key is stored in the database and cookie keys expire, an attacker cannot crack the key in the database within reasonable time. Because cookie keys expire, even stolen login cookies are usable for only a short timespan. And because each site generates their own cookies, cookies cannot be used across sites.